Refactoring dekgl
