+++ superset/assets/visualizations/deckgl/layers/index.js
@@ -3,7 +3,6 @@ import deck_grid from './grid';
 import deck_screengrid from './screengrid';
 import deck_path from './path';
 import deck_hex from './hex';
-import deck_scatter from './scatter';
 import deck_geojson from './geojson';
 import deck_arc from './arc';
 import deck_polygon from './polygon';
@@ -13,7 +12,6 @@ const layerGenerators = {
   deck_screengrid,
   deck_path,
   deck_hex,
-  deck_scatter,
   deck_geojson,
   deck_arc,
   deck_polygon,
